Refactor project structure and add Areas
Updated project structure to include new Admin area with models and view models. Refactored controllers to support area routing and improved method organization. Enhanced views for better styling, user interaction, and validation. Added new functionality for managing attendance, exams, and memorization plans, ensuring a more user-friendly experience. Improved JavaScript search and filtering capabilities across various views.
