+++ AlqudsProject/Areas/Admin/Controllers/MemorizationPathsController.cs
@@ -8,89 +8,89 @@ using Microsoft.EntityFrameworkCore;
 using AlqudsProject.Data;
 using AlqudsProject.Models;
 
-namespace AlqudsProject.Controllers
+namespace AlqudsProject.Areas.Admin.Controllers
 {
-    public class ManhajsController : Controller
+    public class MemorizationPathsController : Controller
     {
         private readonly Alquds_CenterContext _context;
 
-        public ManhajsController(Alquds_CenterContext context)
+        public MemorizationPathsController(Alquds_CenterContext context)
         {
             _context = context;
         }
 
-        // GET: Manhajs
+        // GET: MemorizationPaths
         public async Task<IActionResult> Index()
         {
-              return _context.Manhajs != null ? 
-                          View(await _context.Manhajs.ToListAsync()) :
-                          Problem("Entity set 'Alquds_CenterContext.Manhajs'  is null.");
+              return _context.MemorizationPath != null ? 
+                          View(await _context.MemorizationPath.ToListAsync()) :
+                          Problem("Entity set 'Alquds_CenterContext.MemorizationPath'  is null.");
         }
 
-        // GET: Manhajs/Details/5
+        // GET: MemorizationPaths/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null || _context.Manhajs == null)
+            if (id == null || _context.MemorizationPath == null)
             {
                 return NotFound();
             }
 
-            var manhaj = await _context.Manhajs
-                .FirstOrDefaultAsync(m => m.ManhajId == id);
-            if (manhaj == null)
+            var memorizationPath = await _context.MemorizationPath
+                .FirstOrDefaultAsync(m => m.PathId == id);
+            if (memorizationPath == null)
             {
                 return NotFound();
             }
 
-            return View(manhaj);
+            return View(memorizationPath);
         }
 
-        // GET: Manhajs/Create
+        // GET: MemorizationPaths/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Manhajs/Create
+        // POST: MemorizationPaths/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ManhajId,Name,Url,Picture")] Manhaj manhaj)
+        public async Task<IActionResult> Create([Bind("PathId,MemorizeFrom,MemorizeTo,Name")] MemorizationPath memorizationPath)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(manhaj);
+                _context.Add(memorizationPath);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(manhaj);
+            return View(memorizationPath);
         }
 
-        // GET: Manhajs/Edit/5
+        // GET: MemorizationPaths/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Manhajs == null)
+            if (id == null || _context.MemorizationPath == null)
             {
                 return NotFound();
             }
 
-            var manhaj = await _context.Manhajs.FindAsync(id);
-            if (manhaj == null)
+            var memorizationPath = await _context.MemorizationPath.FindAsync(id);
+            if (memorizationPath == null)
             {
                 return NotFound();
             }
-            return View(manhaj);
+            return View(memorizationPath);
         }
 
-        // POST: Manhajs/Edit/5
+        // POST: MemorizationPaths/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ManhajId,Name,Url,Picture")] Manhaj manhaj)
+        public async Task<IActionResult> Edit(int id, [Bind("PathId,MemorizeFrom,MemorizeTo,Name")] MemorizationPath memorizationPath)
         {
-            if (id != manhaj.ManhajId)
+            if (id != memorizationPath.PathId)
             {
                 return NotFound();
             }
@@ -99,12 +99,12 @@ namespace AlqudsProject.Controllers
             {
                 try
                 {
-                    _context.Update(manhaj);
+                    _context.Update(memorizationPath);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ManhajExists(manhaj.ManhajId))
+                    if (!MemorizationPathExists(memorizationPath.PathId))
                     {
                         return NotFound();
                     }
@@ -115,49 +115,49 @@ namespace AlqudsProject.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            return View(manhaj);
+            return View(memorizationPath);
         }
 
-        // GET: Manhajs/Delete/5
+        // GET: MemorizationPaths/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Manhajs == null)
+            if (id == null || _context.MemorizationPath == null)
             {
                 return NotFound();
             }
 
-            var manhaj = await _context.Manhajs
-                .FirstOrDefaultAsync(m => m.ManhajId == id);
-            if (manhaj == null)
+            var memorizationPath = await _context.MemorizationPath
+                .FirstOrDefaultAsync(m => m.PathId == id);
+            if (memorizationPath == null)
             {
                 return NotFound();
             }
 
-            return View(manhaj);
+            return View(memorizationPath);
         }
 
-        // POST: Manhajs/Delete/5
+        // POST: MemorizationPaths/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Manhajs == null)
+            if (_context.MemorizationPath == null)
             {
-                return Problem("Entity set 'Alquds_CenterContext.Manhajs'  is null.");
+                return Problem("Entity set 'Alquds_CenterContext.MemorizationPath'  is null.");
             }
-            var manhaj = await _context.Manhajs.FindAsync(id);
-            if (manhaj != null)
+            var memorizationPath = await _context.MemorizationPath.FindAsync(id);
+            if (memorizationPath != null)
             {
-                _context.Manhajs.Remove(manhaj);
+                _context.MemorizationPath.Remove(memorizationPath);
             }
             
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ManhajExists(int id)
+        private bool MemorizationPathExists(int id)
         {
-          return (_context.Manhajs?.Any(e => e.ManhajId == id)).GetValueOrDefault();
+          return (_context.MemorizationPath?.Any(e => e.PathId == id)).GetValueOrDefault();
         }
     }
 }
