+++ AlqudsProject/Areas/Admin/Views/Halagat/Delete.cshtml
