+++ AlqudsProject/Areas/Admin/Views/Manhajs/Delete.cshtml
