+++ AlqudsProject/Areas/Admin/Views/Student/Create.cshtml
