+++ AlqudsProject/Areas/Admin/Views/Student/Delete.cshtml
