+++ AlqudsProject/Areas/Admin/Views/Student/Details.cshtml
