+++ AlqudsProject/Areas/Admin/Views/Teacher/Delete.cshtml
