+++ AlqudsProject/Areas/Theachers/Views/Attendance/StudentsDaily.cshtml
@@ -1,4 +1,5 @@
-﻿@model List<StudentDailyRow>
+﻿@using AlqudsProject.ViewModels
+@model List<StudentDailyRow>
 
 @{
     ViewData["Title"] = "التقييم اليومي";
