+++ AlqudsProject/Areas/Theachers/Views/Attendance/TeachersDaily.cshtml
@@ -1,4 +1,5 @@
-﻿@model List<TeacherDailyRow>
+﻿@using AlqudsProject.ViewModels
+@model List<TeacherDailyRow>
 @{
     ViewData["Title"] = "تحضير المعلمات";
     var selectedDate = (DateTime)ViewBag.SelectedDate;
