+++ AlqudsProject/Areas/Theachers/Views/Exams/Add.cshtml
@@ -1,4 +1,5 @@
-﻿@model List<ExamEntryRow>
+﻿@using AlqudsProject.ViewModels
+@model List<ExamEntryRow>
 @{
     ViewData["Title"] = "إضافة اختبار";
     Layout = "~/Views/Shared/_Layout.cshtml";
